refactoring infrastructure
