--- LBC/Services/Base/BaseResult.cs
@@ -1,15 +1,0 @@
-﻿using System;
-namespace LBC.Services.Base
-{
-    public abstract class BaseResult 
-    {
-        public readonly BaseResultStatus.Status Status;
-
-        public BaseResult(BaseResultStatus.Status status)
-        {
-            this.Status = status;
-        }
-
-       
-    }
-}
