--- LBC/Services/Base/BaseResultStatus.cs
@@ -1,14 +1,0 @@
-﻿using System;
-namespace LBC.Services.Base
-{
-    public class BaseResultStatus
-    {
-        public enum Status
-        {
-            Success,
-            Failed,
-            Exception
-        }
-    }
-   
-}
